Fix getcwd size arg to make it work in solaris.
glusterfs/d -f path failed if the path was not absolute.
If the BUF argument is NULL, then on solaris the size argument has to be passed.
On linux, if the size argument passed is 0, then it dynmically allocates a buffer
 of the required size. The fix is to pass PATH_MAX as the size for consistency.

Signed-off-by: Shishir Gowda <shishirng@gluster.com>
Signed-off-by: Anand V. Avati <avati@dev.gluster.com>

BUG: 2443 (starting glusterfs/d -f with relative path crashes)
URL: http://bugs.gluster.com/cgi-bin/bugzilla3/show_bug.cgi?id=2443